Refactor frontend structure
