+++ frontend/src/page/HomePage/Home.jsx
@@ -1,0 +1,11 @@
+function Home () {
+  return (
+    <>
+    
+      <h1>Vite + React</h1>
+     
+    </>
+  )
+}
+
+export default Home;
